requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
langchain>=0.0.350
langchain-openai>=0.0.5
langchain-community>=0.0.15
faiss-cpu>=1.7.4
pypdf>=3.17.1
python-magic>=0.4.27 